+++ docs/releases/releases.md
@@ -44,9 +44,13 @@ The main changes in this release are:
 
 + Change the default value of `encoding` to `ordinal` in <code>[ForecasterAutoregMultiSeries]</code>. This will avoid conflicts if the regressor does not support categorical variables by default.
 
-+ Removed argument `engine` fom <code>[bayesian_search_forecaster]</code> and <code>[bayesian_search_forecaster_multiseries]</code>.
++ Removed argument `engine` from <code>[bayesian_search_forecaster]</code> and <code>[bayesian_search_forecaster_multiseries]</code>.
 
 + The `pmdarima.ARIMA` regressor is no longer supported by the `ForecasterSarimax`. You can use the `skforecast.Sarimax.Sarimax` model or, to continue using it, use skforecast 0.13.0 or lower.
+
++ `initialize_lags` now returns the maximum lag, `max_lag`.
+
++ Removed attribute `window_size_diff` from all Forecasters. The window size extended by the order of differentiation is now calculated on `window_size`.
 
 + Renamed attributes in all Forecasters: